fix: merge AppExtension in CoreExtension
