--- EMS/core-bundle/src/Twig/AppExtension.php
@@ -1,966 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace EMS\CoreBundle\Twig;
-
-use Caxy\HtmlDiff\HtmlDiff;
-use Doctrine\Bundle\DoctrineBundle\Registry;
-use Elastica\ResultSet;
-use EMS\CommonBundle\Common\EMSLink;
-use EMS\CommonBundle\Elasticsearch\Document\DocumentInterface;
-use EMS\CommonBundle\Elasticsearch\Exception\NotFoundException;
-use EMS\CommonBundle\Helper\EmsFields;
-use EMS\CommonBundle\Search\Search as CommonSearch;
-use EMS\CommonBundle\Service\ElasticaService;
-use EMS\CommonBundle\Storage\Processor\Config;
-use EMS\CommonBundle\Storage\Service\StorageInterface;
-use EMS\CommonBundle\Twig\AssetExtension;
-use EMS\CommonBundle\Twig\RequestExtension;
-use EMS\CoreBundle\Core\ContentType\ContentTypeFields;
-use EMS\CoreBundle\Core\ContentType\ContentTypeRoles;
-use EMS\CoreBundle\Entity\ContentType;
-use EMS\CoreBundle\Entity\Environment;
-use EMS\CoreBundle\Entity\FieldType;
-use EMS\CoreBundle\Entity\Sequence;
-use EMS\CoreBundle\Entity\UserInterface;
-use EMS\CoreBundle\Exception\CantBeFinalizedException;
-use EMS\CoreBundle\Exception\SkipNotificationException;
-use EMS\CoreBundle\Form\DataField\DateRangeFieldType;
-use EMS\CoreBundle\Form\DataField\TimeFieldType;
-use EMS\CoreBundle\Form\Factory\ObjectChoiceListFactory;
-use EMS\CoreBundle\Repository\SequenceRepository;
-use EMS\CoreBundle\Routes;
-use EMS\CoreBundle\Service\ContentTypeService;
-use EMS\CoreBundle\Service\FileService;
-use EMS\CoreBundle\Service\Revision\RevisionService;
-use EMS\CoreBundle\Service\SearchService;
-use EMS\CoreBundle\Service\UserService;
-use EMS\Helpers\Standard\Color;
-use EMS\Helpers\Standard\DateTime;
-use EMS\Helpers\Standard\Json;
-use EMS\Helpers\Standard\Type;
-use Symfony\Component\Form\FormError;
-use Symfony\Component\Form\FormFactory;
-use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
-use Symfony\Component\Routing\RouterInterface;
-use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
-use Twig\Environment as TwigEnvironment;
-use Twig\Extension\AbstractExtension;
-use Twig\TwigFilter;
-use Twig\TwigFunction;
-
-class AppExtension extends AbstractExtension
-{
-    /**
-     * @param array<mixed> $assetConfig
-     */
-    public function __construct(
-        private readonly Registry $doctrine,
-        private readonly AuthorizationCheckerInterface $authorizationChecker,
-        private readonly UserService $userService,
-        private readonly RevisionService $revisionService,
-        private readonly ContentTypeService $contentTypeService,
-        private readonly RouterInterface $router,
-        private readonly TwigEnvironment $twig,
-        private readonly ObjectChoiceListFactory $objectChoiceListFactory,
-        protected FormFactory $formFactory,
-        protected FileService $fileService,
-        protected RequestExtension $commonRequestExtension,
-        private readonly ElasticaService $elasticaService,
-        private readonly SearchService $searchService,
-        private readonly AssetExtension $assetExtension,
-        protected array $assetConfig,
-    ) {
-    }
-
-    #[\Override]
-    public function getFunctions(): array
-    {
-        return [
-            new TwigFunction('emsco_cant_be_finalized', $this->cantBeFinalized(...)),
-            new TwigFunction('emsco_sequence', $this->getSequenceNextValue(...)),
-            new TwigFunction('emsco_diff_text', $this->diffText(...), ['is_safe' => ['html']]),
-            new TwigFunction('emsco_diff', $this->diff(...), ['is_safe' => ['html']]),
-            new TwigFunction('emsco_diff_html', $this->diffHtml(...), ['is_safe' => ['html']]),
-            new TwigFunction('emsco_diff_icon', $this->diffIcon(...), ['is_safe' => ['html']]),
-            new TwigFunction('emsco_diff_raw', $this->diffRaw(...), ['is_safe' => ['html']]),
-            new TwigFunction('emsco_diff_color', $this->diffColor(...), ['is_safe' => ['html']]),
-            new TwigFunction('emsco_diff_boolean', $this->diffBoolean(...), ['is_safe' => ['html']]),
-            new TwigFunction('emsco_diff_choice', $this->diffChoice(...), ['is_safe' => ['html']]),
-            new TwigFunction('emsco_diff_data_link', $this->diffDataLink(...), ['is_safe' => ['html']]),
-            new TwigFunction('emsco_diff_date', $this->diffDate(...), ['is_safe' => ['html']]),
-            new TwigFunction('emsco_diff_time', $this->diffTime(...), ['is_safe' => ['html']]),
-            new TwigFunction('emsco_is_super', $this->isSuper(...)),
-            new TwigFunction('emsco_asset_path', $this->assetPath(...), ['is_safe' => ['html']]),
-            new TwigFunction('emsco_call_user_func', $this->callUserFunc(...)),
-            new TwigFunction('emsco_skip_notification', $this->skipNotificationException(...), ['is_safe' => ['html']]),
-            new TwigFunction('emsco_save_contents', $this->saveContents(...)),
-            new TwigFunction('emsco_search', $this->search(...)),
-        ];
-    }
-
-    #[\Override]
-    public function getFilters(): array
-    {
-        return [
-            new TwigFilter('emsco_convert_java_date_format', fn (string $format) => DateTime::convertFormat('java', $format)),
-            new TwigFilter('emsco_convert_javascript_date_format', fn (string $format) => DateTime::convertFormat('js', $format)),
-            new TwigFilter('emsco_convert_javascript_date_range_format', $this->convertJavascriptDateRangeFormat(...)),
-            new TwigFilter('emsco_time_field_time_format', $this->getTimeFieldTimeFormat(...)),
-            new TwigFilter('emsco_soap_request', $this->soapRequest(...)),
-            new TwigFilter('emsco_all_granted', $this->allGranted(...)),
-            new TwigFilter('emsco_one_granted', $this->oneGranted(...)),
-            new TwigFilter('emsco_in_my_circles', $this->inMyCircles(...)),
-            new TwigFilter('emsco_data_link', $this->dataLink(...), ['is_safe' => ['html']]),
-            new TwigFilter('emsco_generate_from_template', $this->generateFromTemplate(...)),
-            new TwigFilter('emsco_object_choice_loader', $this->objectChoiceLoader(...)),
-            new TwigFilter('emsco_grouped_object_loader', $this->groupedObjectLoader(...)),
-            new TwigFilter('emsco_property_path', $this->propertyPath(...)),
-            new TwigFilter('emsco_internal_links', $this->internalLinks(...)),
-            new TwigFilter('emsco_get_user', $this->getUser(...)),
-            new TwigFilter('emsco_display_name', $this->displayName(...)),
-            new TwigFilter('emsco_date_difference', $this->dateDifference(...)),
-            new TwigFilter('emsco_src_path', $this->srcPath(...)),
-            new TwigFilter('emsco_search', $this->search(...)),
-            new TwigFilter('emsco_search_query', $this->searchQuery(...)),
-            new TwigFilter('emsco_call_user_func', $this->callUserFunc(...)),
-            new TwigFilter('emsco_get_string', $this->getString(...)),
-            new TwigFilter('emsco_get_field_by_path', $this->getFieldByPath(...)),
-            new TwigFilter('emsco_get', $this->get(...)),
-            new TwigFilter('emsco_get_file', $this->getFile(...)),
-        ];
-    }
-
-    /**
-     * @param array<mixed> $fileField
-     * @param array<mixed> $assetConfig
-     * @param int          $referenceType
-     */
-    public function assetPath(array $fileField, string $processorIdentifier, array $assetConfig = [], string $route = 'ems_asset', string $fileHashField = EmsFields::CONTENT_FILE_HASH_FIELD, string $filenameField = EmsFields::CONTENT_FILE_NAME_FIELD, string $mimeTypeField = EmsFields::CONTENT_MIME_TYPE_FIELD, $referenceType = UrlGeneratorInterface::RELATIVE_PATH): string
-    {
-        $config = $assetConfig;
-        if (!isset($config['_config_type'])) {
-            $config['_config_type'] = 'image';
-        }
-
-        if (isset($this->assetConfig[$processorIdentifier])) {
-            $config = \array_merge($this->assetConfig[$processorIdentifier], $config);
-        }
-
-        // removes invalid options like _sha1, _finalized_by, ..
-        $config = \array_intersect_key($config, Config::getDefaults());
-
-        // _published_datetime can also be removed as it has a sense only if the default config is updated
-        if (isset($config['_published_datetime'])) {
-            unset($config['_published_datetime']);
-        }
-
-        return $this->assetExtension->assetPath($fileField, $assetConfig, $route, $fileHashField, $filenameField, $mimeTypeField, $referenceType);
-    }
-
-    public function getFieldByPath(ContentType $contentType, string $path, bool $skipVirtualFields = false): ?FieldType
-    {
-        $fieldType = $this->contentTypeService->getChildByPath($contentType->getFieldType(), $path, $skipVirtualFields);
-        if (false === $fieldType) {
-            return null;
-        }
-
-        return $fieldType;
-    }
-
-    public function getFile(string $hash): ?string
-    {
-        return $this->fileService->getFile($hash);
-    }
-
-    /**
-     * @return mixed[]
-     */
-    public function saveContents(string $content, string $filename, string $mimetype, int $usage = StorageInterface::STORAGE_USAGE_ASSET): array
-    {
-        $hash = $this->fileService->saveContents($content, $filename, $mimetype, $usage);
-
-        return [
-            EmsFields::CONTENT_FILE_HASH_FIELD => $hash,
-            EmsFields::CONTENT_FILE_SIZE_FIELD => \strlen($content),
-            EmsFields::CONTENT_FILE_NAME_FIELD => $filename,
-            EmsFields::CONTENT_MIME_TYPE_FIELD => $mimetype,
-        ];
-    }
-
-    /**
-     * @param array<mixed> $rawData
-     */
-    public function getString(array $rawData, string $field): ?string
-    {
-        if (empty($rawData) or !isset($rawData[$field])) {
-            return null;
-        }
-        if (\is_string($rawData[$field])) {
-            return $rawData[$field];
-        }
-        $encoded = Json::encode($rawData[$field]);
-
-        return $encoded;
-    }
-
-    public function diff(?string $a, ?string $b, bool $compare, bool $escape = false, bool $htmlDiff = false, bool $raw = false): string
-    {
-        $tag = 'span';
-        $textClass = '';
-        $textLabel = '';
-
-        if ($compare && $a !== $b) {
-            if ($htmlDiff && $a && $b) {
-                $textClass = 'text-orange';
-                $htmlDiff = new HtmlDiff(($escape ? \htmlentities($b) : $this->internalLinks($b)) ?? '', ($escape ? \htmlentities($a) : $this->internalLinks($a)) ?? '');
-                $textLabel = $htmlDiff->build();
-            } else {
-                $textClass = false;
-                if (null !== $b) {
-                    $textClass = 'text-red';
-                    $textLabel .= '<del class="diffmod">'.($escape ? \htmlentities($b) : $this->internalLinks($b)).'</del>';
-                }
-
-                if (null !== $a) {
-                    if ($textClass) {
-                        $textClass = 'text-orange';
-                    } else {
-                        $textClass = 'text-green';
-                    }
-                    $textLabel .= ' <ins class="diffmod">'.($escape ? \htmlentities($a) : $this->internalLinks($a)).'</ins>';
-                }
-            }
-        } else {
-            if (null !== $a) {
-                $textLabel = ($escape ? \htmlentities($a) : $this->internalLinks($a));
-            } else {
-                return '<span class="text-gray">[not defined]</span>';
-            }
-        }
-
-        if ($raw) {
-            return $textLabel ?? '';
-        }
-
-        return '<'.$tag.' class="'.$textClass.'">'.$textLabel.'</'.$tag.'>';
-    }
-
-    /**
-     * @param mixed|null $rawData
-     * @param mixed|null $compareRawData
-     */
-    public function diffBoolean($rawData, bool $compare, string $fieldName, $compareRawData): string
-    {
-        $a = $rawData ? true : false;
-        $b = isset($compareRawData[$fieldName]) && $compareRawData[$fieldName];
-
-        $textClass = '';
-        if ($compare && $a !== $b) {
-            $textClass = 'text-orange';
-        }
-
-        return '<span class="'.$textClass.'"><i class="fa fa'.($a ? '-check' : '').'-square-o"></i></span>';
-    }
-
-    /**
-     * @param mixed[]|null $compareRawData
-     */
-    public function diffIcon(?string $rawData, bool $compare, string $fieldName, ?array $compareRawData): string
-    {
-        $b = $a = null;
-        if (null !== $rawData) {
-            $a = \sprintf('<i class="%s"></i> %s', $rawData, $rawData);
-        }
-
-        $compareData = Type::getAsNullableString($compareRawData[$fieldName] ?? null);
-        if (null !== $compareData) {
-            $b = \sprintf('<i class="%s"></i> %s', $compareData, $compareData);
-        }
-
-        return $this->diff($a, $b, $compare);
-    }
-
-    /**
-     * @param mixed|null $rawData
-     * @param mixed|null $compareRawData
-     */
-    public function diffTime($rawData, bool $compare, string $fieldName, $compareRawData, string $format1, string $format2): string
-    {
-        return $this->diffDate($rawData, $compare, $fieldName, $compareRawData, $format1, $format2, TimeFieldType::STOREFORMAT);
-    }
-
-    /**
-     * @param mixed|null $rawData
-     * @param mixed|null $compareRawData
-     */
-    public function diffDate($rawData, bool $compare, string $fieldName, $compareRawData, string $format1, ?string $format2 = null, ?string $internalFormat = null): string
-    {
-        $b = $a = [];
-        $out = '';
-        $tag = 'li';
-        $insColor = 'green';
-        $delColor = 'red';
-
-        if (isset($compareRawData[$fieldName])) {
-            if (\is_array($compareRawData[$fieldName])) {
-                $b = $compareRawData[$fieldName];
-            } elseif (\is_scalar($compareRawData[$fieldName])) {
-                $b = [$compareRawData[$fieldName]];
-            }
-        }
-
-        if (\is_array($rawData)) {
-            $a = $rawData;
-        } elseif (\is_scalar($rawData)) {
-            $tag = 'span';
-            if (!empty($b)) {
-                $insColor = $delColor = 'orange';
-            }
-            $a = [$rawData];
-        }
-
-        $formatedA = [];
-
-        foreach ($a as $item) {
-            try {
-                if ($item instanceof \DateTime) {
-                    $date = $item;
-                } elseif ($internalFormat) {
-                    $date = \DateTime::createFromFormat($internalFormat, $item);
-                } else {
-                    $date = new \DateTime($item);
-                }
-            } catch (\Throwable) {
-                $date = null;
-            }
-
-            if (!$date instanceof \DateTimeInterface) {
-                $out .= '<'.$tag.' class="text-red">'.\htmlentities((string) $item).'</'.$tag.'>';
-                continue;
-            }
-
-            $value = $date->format($format1);
-            $value2 = null;
-
-            if (null !== $internalFormat) {
-                $internal = $date->format($internalFormat);
-                $formatedA[] = $internal;
-                $inArray = \in_array($internal, $b);
-            } elseif (null !== $format2) {
-                $value2 = $date->format($format2);
-                $formatedA[] = $value2;
-                $inArray = \in_array($item, $b);
-            } else {
-                $formatedA[] = $value;
-                $inArray = \in_array($value, $b);
-            }
-
-            if ($value2) {
-                $value .= ' ('.$value2.')';
-            }
-
-            if (!$compare || $inArray) {
-                $out .= '<'.$tag.' class="">'.\htmlentities($value).'</'.$tag.'>';
-            } else {
-                $out .= '<'.$tag.' class="text-'.$insColor.'"><ins class="diffmod">'.\htmlentities($value).'</ins></'.$tag.'>';
-            }
-        }
-
-        if ($compare) {
-            foreach ($b as $item) {
-                if ($item instanceof \DateTime) {
-                    $date = $item;
-                } elseif ($internalFormat) {
-                    $date = \DateTime::createFromFormat($internalFormat, $item);
-                } else {
-                    $date = new \DateTime($item);
-                }
-                if (false === $date) {
-                    throw new \RuntimeException('Unexpected date format');
-                }
-
-                $value = $date->format($format1);
-                $value2 = null;
-
-                if (null !== $internalFormat) {
-                    $internal = $date->format($internalFormat);
-                    $inArray = \in_array($internal, $formatedA);
-                } elseif (null !== $format2) {
-                    $value2 = $date->format($format2);
-                    $inArray = \in_array($item, $formatedA);
-                } else {
-                    $inArray = \in_array($value, $formatedA);
-                }
-
-                if ($value2) {
-                    $value .= ' ('.$value2.')';
-                }
-
-                if (!$inArray) {
-                    $out .= ' <'.$tag.' class="text-'.$delColor.'"><del class="diffmod">'.\htmlentities($value).'</del></'.$tag.'>';
-                }
-            }
-        }
-
-        return $out;
-    }
-
-    /**
-     * @param mixed|null        $rawData
-     * @param array<mixed>|null $labels
-     * @param array<mixed>|null $choices
-     * @param mixed|null        $compareRawData
-     */
-    public function diffChoice($rawData, ?array $labels, ?array $choices, bool $compare, string $fieldName, $compareRawData): string
-    {
-        $b = $a = [];
-        $out = '';
-        $tag = 'li';
-        $insColor = 'green';
-        $delColor = 'red';
-
-        if (isset($compareRawData[$fieldName])) {
-            if (\is_array($compareRawData[$fieldName])) {
-                $b = $compareRawData[$fieldName];
-            } elseif (\is_scalar($compareRawData[$fieldName])) {
-                $b = [$compareRawData[$fieldName]];
-            }
-        }
-
-        if (\is_array($rawData)) {
-            $a = $rawData;
-        } elseif (\is_scalar($rawData)) {
-            $tag = 'span';
-            if (!empty($b)) {
-                $insColor = $delColor = 'orange';
-            }
-            $a = [$rawData];
-        }
-
-        if ($compare) {
-            foreach ($b as $item) {
-                $value = $item;
-                if (\is_array($choices) && \in_array($value, $choices)) {
-                    $idx = \array_search($value, $choices, true);
-                    if (false !== $idx && \is_array($labels) && \array_key_exists($idx, $labels)) {
-                        $value = $labels[$idx].' ('.$value.')';
-                    }
-                }
-                if (!\in_array($item, $a)) {
-                    $out .= '<'.$tag.' class="text-'.$delColor.'"><del class="diffmod">'.\htmlentities((string) $value).'</del></'.$tag.'>';
-                }
-            }
-        }
-
-        foreach ($a as $item) {
-            $value = $item;
-            if (\is_array($choices) && \in_array($value, $choices)) {
-                $idx = \array_search($value, $choices, true);
-                if (false !== $idx && \is_array($labels) && \array_key_exists($idx, $labels)) {
-                    $value = $this->isSuper() ? $labels[$idx].' ('.$item.')' : $labels[$idx];
-                }
-            }
-            if (!$compare || \in_array($item, $b)) {
-                $out .= '<'.$tag.' class="" data-ems-id="'.$item.'">'.\htmlentities((string) $value).'</'.$tag.'>';
-            } else {
-                $out .= '<'.$tag.' class="text-'.$insColor.'"><ins class="diffmod">'.\htmlentities((string) $value).'</ins></'.$tag.'>';
-            }
-        }
-
-        if (empty($out)) {
-            $out = '<span class="text-gray">[empty]</span>';
-        }
-
-        return $out;
-    }
-
-    /**
-     * @param mixed|null $rawData
-     * @param mixed|null $compareRawData
-     */
-    public function diffDataLink($rawData, bool $compare, string $fieldName, $compareRawData): string
-    {
-        $b = $a = [];
-        $out = '';
-
-        if (\is_array($rawData)) {
-            $a = $rawData;
-        } elseif (\is_scalar($rawData)) {
-            $a = [$rawData];
-        }
-
-        if (isset($compareRawData[$fieldName])) {
-            if (\is_array($compareRawData[$fieldName])) {
-                $b = $compareRawData[$fieldName];
-            } elseif (\is_scalar($compareRawData[$fieldName])) {
-                $b = [$compareRawData[$fieldName]];
-            }
-        }
-
-        if ($compare) {
-            foreach ($b as $item) {
-                if (!\in_array($item, $a)) {
-                    $out .= $this->dataLink($item, null, 'del').' ';
-                }
-            }
-        }
-
-        foreach ($a as $item) {
-            if (!$compare || \in_array($item, $b)) {
-                $out .= $this->dataLink($item).' ';
-            } else {
-                $out .= $this->dataLink($item, null, 'ins').' ';
-            }
-        }
-
-        return $out;
-    }
-
-    /**
-     * @param mixed[]|null $compareRawData
-     */
-    public function diffColor(?string $rawData, bool $compare, string $fieldName, ?array $compareRawData): string
-    {
-        $b = $a = null;
-        if (null !== $rawData) {
-            $color = new Color($rawData);
-            $a = \sprintf('<span style="background-color: %s; color: %s;">%s</span> ', $rawData, $color->bestContrast(...Color::EMS_COLORS)->getRGB(), $rawData);
-        }
-
-        $compareData = Type::getAsNullableString($compareRawData[$fieldName] ?? null);
-        if (null !== $compareData) {
-            $color = new Color($compareData);
-            $b = \sprintf('<span style="background-color: %s; color: %s;">%s</span> ', $compareData, $color->bestContrast(...Color::EMS_COLORS)->getRGB(), $compareData);
-        }
-
-        return $this->diff($a, $b, $compare, false, false, true);
-    }
-
-    /**
-     * @param mixed|null $rawData
-     * @param mixed|null $compareRawData
-     */
-    public function diffRaw($rawData, bool $compare, string $fieldName, $compareRawData): string
-    {
-        if (\is_array($rawData)) {
-            $a = Json::encode($rawData);
-        } else {
-            $a = Type::getAsNullableString($rawData);
-        }
-        $b = $compareRawData[$fieldName] ?? null;
-        if (\is_array($b)) {
-            $b = Json::encode($b);
-        } else {
-            $b = Type::getAsNullableString($b);
-        }
-
-        return $this->diff($a, $b, $compare);
-    }
-
-    /**
-     * @param mixed|null   $rawData
-     * @param mixed[]|null $compareRawData
-     */
-    public function diffText($rawData, bool $compare, string $fieldName, ?array $compareRawData): string
-    {
-        $b = Type::getAsNullableString($compareRawData[$fieldName] ?? null);
-
-        return $this->diff(Type::getAsNullableString($rawData), $b, $compare, true, true);
-    }
-
-    /**
-     * @param mixed|null   $rawData
-     * @param mixed[]|null $compareRawData
-     */
-    public function diffHtml($rawData, bool $compare, string $fieldName, ?array $compareRawData): string
-    {
-        $b = Type::getAsNullableString($compareRawData[$fieldName] ?? null);
-
-        return $this->diff(Type::getAsNullableString($rawData), $b, $compare, false, true, true);
-    }
-
-    public function getSequenceNextValue(string $name): int
-    {
-        $em = $this->doctrine->getManager();
-        $repo = $em->getRepository(Sequence::class);
-        if (!$repo instanceof SequenceRepository) {
-            throw new \RuntimeException('Unexpected repository');
-        }
-
-        return $repo->nextValue($name);
-    }
-
-    public function cantBeFinalized(string $message = '', int $code = 0, ?\Throwable $previous = null): never
-    {
-        throw new CantBeFinalizedException($message, $code, $previous);
-    }
-
-    public function callUserFunc(mixed $function, mixed ...$parameter): mixed
-    {
-        return \call_user_func($function, $parameter);
-    }
-
-    /**
-     * @param array<mixed> $params
-     *
-     * @return array<mixed>
-     */
-    public function searchQuery(array $params): array
-    {
-        $search = $this->elasticaService->convertElasticsearchSearch($params);
-
-        return $this->elasticaService->search($search)->getResponse()->getData();
-    }
-
-    /**
-     * @param string|string[]     $indexes
-     * @param string|array<mixed> $body
-     * @param string|list<string> $contentTypes
-     * @param array<mixed>|null   $sort
-     * @param string[]|null       $sources
-     */
-    public function search(string|array $indexes, string|array $body = [], string|array $contentTypes = [], ?int $size = null, int $from = 0, ?array $sort = null, ?array $sources = null): ResultSet
-    {
-        if (\is_string($contentTypes)) {
-            $contentTypes = [$contentTypes];
-        }
-        $query = $this->elasticaService->filterByContentTypes(null, $contentTypes);
-
-        if (\is_string($body)) {
-            $body = Json::decode($body);
-        }
-        $boolQuery = $this->elasticaService->getBoolQuery();
-        if (!empty($body) && $query instanceof $boolQuery) {
-            $query->addMust($body);
-        } elseif (!empty($body)) {
-            if (null !== $query) {
-                $boolQuery->addMust($query);
-            }
-            $query = $boolQuery;
-            $query->addMust($body);
-        }
-        $search = new CommonSearch(\is_array($indexes) ? $indexes : [$indexes], $query);
-        if (null !== $size) {
-            $search->setSize($size);
-        }
-        $search->setFrom($from);
-        if (null !== $sort) {
-            $search->setSort($sort);
-        }
-        if (null !== $sources) {
-            $search->setSources($sources);
-        }
-
-        return $this->elasticaService->search($search);
-    }
-
-    public function dateDifference(string $date1, string $date2, bool $detailed = false): string
-    {
-        $datetime1 = \date_create($date1);
-        $datetime2 = \date_create($date2);
-
-        if (false === $datetime1 || false === $datetime2) {
-            throw new \RuntimeException('Unexpected date format');
-        }
-
-        $interval = \date_diff($datetime1, $datetime2);
-        if ($detailed) {
-            return $interval->format('%R%a days %h hours %i minutes');
-        }
-
-        return ((int) $interval->format('%R%a') + 1).' days';
-    }
-
-    public function getUser(string $username): ?UserInterface
-    {
-        return $this->userService->getUser($username);
-    }
-
-    public function displayName(?string $username): string
-    {
-        return match ($username) {
-            null, '' => 'N/A',
-            default => $this->userService->searchUser($username)?->getDisplayName() ?? $username,
-        };
-    }
-
-    public function srcPath(string $input, bool $asFileName = false): ?string
-    {
-        $path = $this->router->generate('ems_file_view', ['sha1' => '__SHA1__'], UrlGeneratorInterface::ABSOLUTE_PATH);
-        $path = \substr($path, 0, \strlen($path) - 8);
-
-        $out = \preg_replace_callback(
-            '/(ems:\/\/asset:)(?P<hash>[^\n\r"\'\?]*)(?:\?(?P<query>(?:[^\n\r"|\']*)))?/i',
-            function ($matches) use ($path, $asFileName) {
-                if ($asFileName) {
-                    return $this->fileService->getFile($matches['hash']) ?? $path.$matches['hash'];
-                }
-
-                $parameters = [];
-                $query = \html_entity_decode($matches['query'] ?? '');
-                \parse_str($query, $parameters);
-                if (\is_string($parameters['name'] ?? null) && \is_string($parameters['type'] ?? null)) {
-                    return $this->assetExtension->assetPath(
-                        [
-                            EmsFields::CONTENT_FILE_HASH_FIELD => $matches['hash'],
-                            EmsFields::CONTENT_FILE_NAME_FIELD => $parameters['name'],
-                            EmsFields::CONTENT_MIME_TYPE_FIELD => $parameters['type'],
-                        ],
-                        [
-                        ],
-                        'ems_asset',
-                        EmsFields::CONTENT_FILE_HASH_FIELD,
-                        EmsFields::CONTENT_FILE_NAME_FIELD,
-                        EmsFields::CONTENT_MIME_TYPE_FIELD,
-                        UrlGeneratorInterface::ABSOLUTE_PATH
-                    );
-                }
-
-                return $path.$matches['hash'];
-            },
-            $input
-        );
-        $path = $this->router->generate(Routes::DATA_LINK, ['key' => '__KEY__'], UrlGeneratorInterface::ABSOLUTE_PATH);
-        $out = \preg_replace_callback(
-            '/ems:\/\/(?P<key>file:([^\n\r"\'\?]*))/i',
-            fn ($matches) => \str_replace('__KEY__', $matches['key'], $path),
-            Type::string($out)
-        );
-
-        return $out;
-    }
-
-    public function internalLinks(string $input, bool $asFileName = false): ?string
-    {
-        $url = $this->router->generate(Routes::DATA_LINK, ['key' => 'object:'], UrlGeneratorInterface::ABSOLUTE_PATH);
-        $out = \preg_replace('/ems:\/\/object:/i', $url, $input);
-
-        if (null === $out) {
-            throw new \RuntimeException('Unexpected null value');
-        }
-
-        return $this->srcPath($out, $asFileName);
-    }
-
-    public function isSuper(): bool
-    {
-        return $this->userService->isSuper();
-    }
-
-    /**
-     * @param string[] $roles
-     */
-    public function allGranted(array $roles, bool $super = false): bool
-    {
-        if ($super && !$this->isSuper()) {
-            return false;
-        }
-
-        return \array_all($roles, fn ($role) => $this->authorizationChecker->isGranted($role));
-    }
-
-    /**
-     * @param string|string[] $circles
-     */
-    public function inMyCircles(string|array $circles): bool
-    {
-        return $this->userService->inMyCircles($circles);
-    }
-
-    /**
-     * @return array<string>
-     */
-    public function objectChoiceLoader(string $contentTypeName): array
-    {
-        return $this->objectChoiceListFactory->createLoader($contentTypeName, true)->loadAll();
-    }
-
-    /**
-     * @return array<int|string, array<int, mixed>>
-     */
-    public function groupedObjectLoader(string $contentTypeName): array
-    {
-        $choices = $this->objectChoiceListFactory->createLoader($contentTypeName, true)->loadAll();
-        $out = [];
-        foreach ($choices as $choice) {
-            if (!isset($out[$choice->getGroup()])) {
-                $out[$choice->getGroup()] = [];
-            }
-            $out[$choice->getGroup()][] = $choice;
-        }
-
-        return $out;
-    }
-
-    /**
-     * @param array<mixed> $params
-     */
-    public function generateFromTemplate(?string $template, array $params = []): ?string
-    {
-        if (empty($template)) {
-            return null;
-        }
-        try {
-            $out = $this->twig->createTemplate($template)->render($params);
-        } catch (\Exception $e) {
-            $out = 'Error in template: '.$e->getMessage();
-        }
-
-        return $out;
-    }
-
-    public function dataLink(string $key, ?string $revisionId = null, ?string $diffMod = null): string
-    {
-        $emsLink = EMSLink::fromText($key);
-        if (!$emsLink->isValid() || !$contentType = $this->contentTypeService->getByName($emsLink->getContentType())) {
-            return $key;
-        }
-
-        $label = \sprintf('<i class="%s"></i>', $contentType->getIcon() ?? 'fa fa-book');
-
-        try {
-            $document = $this->searchService->getDocument($contentType, $emsLink->getOuuid());
-            $emsLink = $document->getEmsLink(); // versioned documents
-            $emsSource = $document->getEMSSource();
-            $label .= \sprintf('<span>%s</span>', \htmlentities($this->revisionService->display($document)));
-        } catch (NotFoundException) {
-            $label .= \sprintf('<span>%s</span>', $emsLink->getEmsId());
-        }
-
-        $color = isset($emsSource) && $contentType->hasColorField() ? $emsSource->get($contentType->giveColorField()) : null;
-        if ($color) {
-            $contrasted = $this->contrastRatio($color, '#000000') > $this->contrastRatio($color, '#ffffff') ? '#000000' : '#ffffff';
-            $label = '<span class="" style="color:'.$contrasted.';">'.$label.'</span>';
-        }
-
-        $attributes = [];
-        $out = $label;
-
-        if (null !== $diffMod) {
-            $out = '<'.$diffMod.' class="diffmod">'.$out.'<'.$diffMod.'>';
-        }
-
-        $tooltipField = $contentType->field(ContentTypeFields::TOOLTIP);
-        if ($tooltipField && isset($emsSource) && $tooltip = $emsSource->get($tooltipField, false)) {
-            $attributes = ['data-toggle="tooltip"', 'data-placement="top"', \sprintf('title="%s"', $tooltip)];
-        }
-
-        if (!$this->authorizationChecker->isGranted($contentType->role(ContentTypeRoles::VIEW))) {
-            if ($color) {
-                $attributes = [...$attributes, \sprintf('style="color: %s"', $color)];
-            }
-
-            return \sprintf('<span %s>%s</span>', \implode(' ', $attributes), $out);
-        }
-
-        if (isset($color)) {
-            $attributes = [...$attributes, 'style="background-color: '.$color.';border-color: '.$color.';"'];
-        }
-
-        $link = $this->router->generate('emsco_view_revisions', [
-            'type' => $emsLink->getContentType(),
-            'ouuid' => $emsLink->getOuuid(),
-            'revisionId' => $revisionId,
-        ]);
-
-        return \vsprintf(
-            '<a class="ems-data-link btn btn-primary btn-sm" href="%s" %s>%s</a>',
-            [$link, \implode(' ', $attributes), $out]
-        );
-    }
-
-    public static function propertyPath(FormError $error): string
-    {
-        $parent = $error->getOrigin();
-        $out = '';
-        while ($parent) {
-            $out = $parent->getName().$out;
-            $parent = $parent->getParent();
-            if ($parent) {
-                $out = '_'.$out;
-            }
-        }
-
-        return $out;
-    }
-
-    public function get(?string $key, ?Environment $environment = null): ?DocumentInterface
-    {
-        if (empty($key)) {
-            return null;
-        }
-
-        $exploded = \explode(':', $key);
-        if (2 !== \count($exploded)) {
-            return null;
-        }
-        $type = $exploded[0];
-        $ouuid = $exploded[1];
-
-        $contentType = $this->contentTypeService->getByName($type);
-        if (!$contentType instanceof ContentType) {
-            return null;
-        }
-
-        try {
-            return $this->searchService->getDocument($contentType, $ouuid, $environment);
-        } catch (NotFoundException) {
-            return null;
-        }
-    }
-
-    /**
-     * @param string[] $roles
-     */
-    public function oneGranted(array $roles, bool $super = false): bool
-    {
-        if ($super && !$this->isSuper()) {
-            return false;
-        }
-
-        return \array_any($roles, fn ($role) => $this->authorizationChecker->isGranted($role));
-    }
-
-    private function contrastRatio(string $c1, string $c2): float
-    {
-        $color1 = new Color($c1);
-        $color2 = new Color($c2);
-
-        return $color1->contrastRatio($color2);
-    }
-
-    public function convertJavascriptDateRangeFormat(string $format): string
-    {
-        return DateRangeFieldType::convertJavascriptDateRangeFormat($format);
-    }
-
-    /**
-     * @param array<array<string>> $options
-     */
-    public function getTimeFieldTimeFormat(array $options): string
-    {
-        return TimeFieldType::getFormat($options);
-    }
-
-    /**
-     * @param array{function: string, options?: array<mixed>, parameters?: mixed} $arguments
-     */
-    public function soapRequest(mixed $wsdl, array $arguments): mixed
-    {
-        $soapClient = new \SoapClient($wsdl, $arguments['options'] ?? []);
-        $function = $arguments['function'];
-
-        if (\array_key_exists('parameters', $arguments)) {
-            return $soapClient->$function($arguments['parameters']);
-        }
-
-        return $soapClient->$function();
-    }
-
-    public function skipNotificationException(string $message = 'This notification has been skipped'): never
-    {
-        throw new SkipNotificationException($message);
-    }
-}
